Support nonoverlapping reasoning of 128-bit loads and other access bitwidths
This patch adds support for nonoverlapping reasoning of 128-bit loads and other access bitwidths.
Also, it adds `bytes128` that is analogous to `bytes64`.
With this, the specification does not need to rely on the low-level `wbytes` definition anymore.

This patch allows s2n-bignum to prove this cute lemma which simply states that `stp` does not touch the values stored at `[x0]`:
https://gist.github.com/aqjune/3c72fbce77f04eefb97a83ca30cef361

Minor updates were necessary in a few proofs in x86/proofs/ because the changes in `READ_MEMORY_BYTESIZED_SPLIT` affected the proofs.

Also, this adds `.gitignore` that lists `.o`, `.obj`, `.correct` and `.a`. This hides the compiled binaries and previously proven facts from the list printed by `git status`.