Add simple generic modular exponentiation
The "bignum_modexp" function is a simple generic function for modular
exponentiation (with odd modulus). Because it is a relatively slow
operation, this update also slightly generalizes the benchmarking setup
so that some functions (including this one) are repeated fewer times,
making the full benchmarking run complete in a reasonable time.